HW5 refactor tests, added tests for Map- and ListStorage

+++ src/com/petrovvich/webapp/storage/ListStorage.java
@@ -31,7 +31,7 @@ public class ListStorage extends AbstractStorage {
 
     @Override
     protected void deleteResumeFromStorage(Object searchIndex) {
-        storage.remove(searchIndex);
+        storage.remove(((Integer) searchIndex).intValue());
     }
 
     @Override
